+++ kamon-akka-http/src/main/resources/reference.conf
@@ -28,8 +28,7 @@ kamon.akka-http {
 
 kanela {
   show-banner = true
-  debug-mode = true
-  log-level = "INFO"
+  log-level = "ERROR"
   modules {
     akka-http-module {
       name = "Akka-Http Instrumentation Module"
