+++ kamon-akka-http/src/main/scala/kamon/akka/http/instrumentation/kanela/mixin/HasContextMixin.scala
@@ -14,17 +14,16 @@
  * =========================================================================================
  */
 
-package akka.http.impl.engine.client
+package kamon.akka.http.instrumentation.kanela.mixin
 
-import akka.http.impl.engine.client.PoolInterfaceActor.PoolRequest
-import akka.http.scaladsl.model.headers.RawHeader
 import kamon.Kamon
-import kamon.context.HasContext
+import kamon.context.{Context, HasContext}
+import kanela.agent.api.instrumentation.mixin.Initializer
 
-object PoolRequestInstrumentation {
-    def attachContext(poolRequest: PoolRequest): AnyRef = {
-      val contextHeaders = Kamon.contextCodec().HttpHeaders.encode(poolRequest.asInstanceOf[HasContext].context).values.map(c => RawHeader(c._1, c._2))
-      val requestWithContext = poolRequest.request.withHeaders(poolRequest.request.headers ++ contextHeaders)
-      poolRequest.copy(request = requestWithContext)
-    }
+class HasContextMixin extends HasContext {
+  var context: Context = _
+
+  @Initializer
+  def _initializer(): Unit = this.context =
+    Kamon.currentContext()
 }
